requirements.txt
pydantic~=2.4.2
mutagen~=1.47.0
rich~=13.6.0
requests~=2.31.0
python-dotenv
git+https://github.com/panos-stavrianos/DirectusPyWrapper.git